+++ feature/evaluate/src/main/java/com/my/version/feature/evaluate/first/navigation/EvaluationFirstNavigation.kt
@@ -1,4 +1,4 @@
-package com.my.version.feature.evaluate.navigation
+package com.my.version.feature.evaluate.first.navigation
 
 import androidx.compose.ui.Modifier
 import androidx.navigation.NavController
@@ -6,7 +6,8 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import com.my.version.core.common.navigation.Route
-import com.my.version.feature.evaluate.EvaluationFirstRoute
+import com.my.version.feature.evaluate.first.EvaluationFirstRoute
+import com.my.version.feature.evaluate.main.navigation.Evaluation
 import kotlinx.serialization.Serializable
 
 fun NavController.navigateToEvaluationFirst(navOptions: NavOptions) = navigate(Evaluation, navOptions)
