+++ feature/evaluate/src/main/java/com/my/version/feature/evaluate/main/EvaluationViewModel.kt
@@ -1,4 +1,4 @@
-package com.my.version.feature.evaluate
+package com.my.version.feature.evaluate.main
 
 import androidx.lifecycle.ViewModel
 import javax.inject.Inject
