+++ feature/evaluate/src/main/java/com/my/version/feature/evaluate/main/navigation/EvaluationNavigation.kt
@@ -1,4 +1,4 @@
-package com.my.version.feature.evaluate.navigation
+package com.my.version.feature.evaluate.main.navigation
 
 import androidx.compose.ui.Modifier
 import androidx.navigation.NavController
@@ -6,7 +6,7 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import com.my.version.core.common.navigation.MainTabRoute
-import com.my.version.feature.evaluate.EvaluationRoute
+import com.my.version.feature.evaluate.main.EvaluationRoute
 import kotlinx.serialization.Serializable
 
 fun NavController.navigateToEvaluation(navOptions: NavOptions) = navigate(Evaluation, navOptions)
