+++ feature/main/src/main/java/com/my/version/feature/main/MainNavigator.kt
@@ -9,11 +9,9 @@ import androidx.navigation.NavHostController
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.navOptions
-import com.my.version.feature.cover.navigation.CoverSecond
 import com.my.version.feature.cover.navigation.navigateToCover
-import com.my.version.feature.evaluate.navigation.Evaluation
-import com.my.version.feature.evaluate.navigation.EvaluationFirst
-import com.my.version.feature.evaluate.navigation.navigateToEvaluation
+import com.my.version.feature.evaluate.first.navigation.EvaluationFirst
+import com.my.version.feature.evaluate.main.navigation.navigateToEvaluation
 import com.my.version.feature.home.navigation.navigateToHome
 
 class MainNavigator(
